arch/x86: port kernel to genuine 32-bit x86 (ARCH=x86)
Adds kernel/arch/x86/ with all 32-bit architecture files:
  boot.S, linker.ld, arch.c, console.c, fb_console.c, serial.c
  interrupts.c, isr.S, syscall_entry.S, user_jump.S,
  context_switch.S, paging.c, lapic.c, io.c, rtc.c,
  signal.c, tlb.c, coredump.c, gdbstub.c, ap_trampoline.S, fpu.S

Core changes:
- Makefile: ARCH=x86 selects i686-elf toolchain, -m32 flags
- mm.h: DIRECT_MAP_BASE/KHEAP_START for 32-bit address space
- user.h: USER_SPACE_LIMIT/USER_STACK_TOP for 32-bit
- sched.h: cpu_context with 32-bit registers (esp/ebp/ebx/esi/edi)
- arch_x86.h: interrupt_frame for 32-bit (u32 registers)
- paging.c: clone identity map into each address space (triple-fault fix)
- process.c: ELF32 loader, PIE_LOAD_BASE=0x40000000 for x86
- scheduler.c: correct 32-bit stack alignment for kthreads
- syscall.c / kheap.c: is_canonical for 32-bit
- userspace/crt/crt0.S: 32-bit _start
- userspace/include/syscall.h: int $0x80 path

Docs: docs/x86-32bit-port.md

+++ kernel/lib/string.c
@@ -7,7 +7,7 @@ void *memcpy(void *dest, const void *src, size_t count)
 	const unsigned char *s = src;
 	
 	while (count >= 8) {
-		*(size_t *)d = *(const size_t *)s;
+		*(unsigned long long *)d = *(const unsigned long long *)s;
 		d += 8;
 		s += 8;
 		count -= 8;
@@ -23,11 +23,11 @@ void *memset(void *dest, int value, size_t count)
 {
 	unsigned char *d = dest;
 	unsigned char v = (unsigned char)value;
-	size_t v64 = ((size_t)v << 56) | ((size_t)v << 48) | ((size_t)v << 40) | ((size_t)v << 32) | 
-	             ((size_t)v << 24) | ((size_t)v << 16) | ((size_t)v << 8) | (size_t)v;
+	unsigned long long v64 = ((unsigned long long)v << 56) | ((unsigned long long)v << 48) | ((unsigned long long)v << 40) | ((unsigned long long)v << 32) | 
+	             ((unsigned long long)v << 24) | ((unsigned long long)v << 16) | ((unsigned long long)v << 8) | (unsigned long long)v;
 
 	while (count >= 8) {
-		*(size_t *)d = v64;
+		*(unsigned long long *)d = v64;
 		d += 8;
 		count -= 8;
 	}
@@ -45,7 +45,7 @@ void *memmove(void *dest, const void *src, size_t count)
 
 	if (d < s) {
 		while (count >= 8) {
-			*(size_t *)d = *(const size_t *)s;
+			*(unsigned long long *)d = *(const unsigned long long *)s;
 			d += 8;
 			s += 8;
 			count -= 8;
@@ -60,7 +60,7 @@ void *memmove(void *dest, const void *src, size_t count)
 		while (count >= 8) {
 			d -= 8;
 			s -= 8;
-			*(size_t *)d = *(const size_t *)s;
+			*(unsigned long long *)d = *(const unsigned long long *)s;
 			count -= 8;
 		}
 		while (count > 0) {
